@@ -1,60 +1,30 @@
 {
-  "name": "findit",
-  "version": "0.1.0",
-  "private": true,
+  "name": "client",
+  "version": "0.0.0",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "prod": " NODE_ENV=production vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview --port 5050",
+    "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs --fix --ignore-path .gitignore"
   },
   "dependencies": {
-    "aws-sdk": "^2.1058.0",
-    "bcrypt": "^5.0.0",
-    "bcryptjs": "^2.4.3",
-    "core-js": "^3.6.5",
     "dotenv": "^16.0.0",
-    "express": "^4.16.1",
-    "ip": "^1.1.5",
-    "mock-aws-s3": "^4.0.2",
-    "nock": "^13.2.2",
-    "request": "^2.79.0",
-    "socket.io": "^4.4.1",
+    "dotenv-webpack": "^7.1.0",
+    "next": "^12.1.0",
+    "socket-io": "^1.0.0",
     "socket.io-client": "^4.4.1",
-    "vue": "^2.6.11",
-    "vue-router": "^3.5.3",
-    "vue-socket.io": "^3.0.10",
-    "vuex": "^3.6.2"
+    "vue": "^3.2.31",
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-plugin-router": "^4.5.15",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "node-sass": "^7.0.1",
-    "sass": "^1.49.0",
-    "vue-template-compiler": "^2.6.11"
-  },
-  "eslintConfig": {
-    "root": true,
-    "env": {
-      "node": true
-    },
-    "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
-    ],
-    "parserOptions": {
-      "parser": "babel-eslint"
-    },
-    "rules": {}
-  },
-  "browserslist": [
-    "> 1%",
-    "last 2 versions",
-    "not dead"
-  ]
+    "@rushstack/eslint-patch": "^1.1.0",
+    "@vitejs/plugin-vue": "^2.2.2",
+    "@vitejs/plugin-vue-jsx": "^1.3.7",
+    "@vue/eslint-config-prettier": "^7.0.0",
+    "eslint": "^8.5.0",
+    "eslint-plugin-vue": "^8.2.0",
+    "prettier": "^2.5.1",
+    "vite": "^2.8.4"
+  }
 }
