@@ -1,17 +1,14 @@
-require("dotenv").config();
-
+const runIn = "development";
 const apiService = {
   //DATA FETCHER
   async fetchData(adress, payload) {
-    console.log("is running in", process.env.NODE_ENV);
     let call;
-    if (process.env.NODE_ENV == "development") {
+    if (runIn == "development") {
       call = "http://192.168.2.100:6080";
     } else {
       call = "https://auth.drunc.net";
     }
     try {
-      console.log("call ", call)
       const request = await fetch(`${call}/${adress}`, {
         body: JSON.stringify(payload),
         headers: {
@@ -34,7 +31,7 @@ const apiService = {
   },
   async projectcall(adress, payload) {
     let call;
-    if (process.env.NODE_ENV == "development") {
+    if (runIn == "development") {
       call = "http://192.168.2.100:7081";
     } else {
       call = "https://api.drunc.net";
