@@ -1,5 +1,9 @@
 <template>
-  <svg viewBox="0 0 8 19" xmlns="http://www.w3.org/2000/svg" preserveAspectRatio="none">
+  <svg
+    viewBox="0 0 8 19"
+    xmlns="http://www.w3.org/2000/svg"
+    preserveAspectRatio="none"
+  >
     <path
       d="M0 15A1 1 0 005 15 1 1 0 000 15M2 13C2 9 2 6 3 3S7-1 7 4C7 9 11 17 2 19"
     />
