@@ -2,8 +2,18 @@
   <div class="projectCluster_wrapper">
     <p v-if="projects.length == 0">Looks quiet empty here</p>
     <div class="projectcluster_preview">
-      <ProjectPreview class="projectcluster_project" :project="{ id: 'addProject' }" @click="goToProject('-1')" v-if="hasAdd" />
-      <div class="projectcluster_project" v-for="proj in projects" :key="proj" @click="goToProject(proj)">
+      <ProjectPreview
+        class="projectcluster_project"
+        :project="{ id: 'addProject' }"
+        @click="goToProject('-1')"
+        v-if="hasAdd"
+      />
+      <div
+        class="projectcluster_project"
+        v-for="proj in projects"
+        :key="proj"
+        @click="goToProject(proj)"
+      >
         <ProjectPreview :project="{ id: proj }" />
         <div v-if="sharedbyself"></div>
         <p>{{ proj }}</p>
