@@ -9,15 +9,27 @@
 
     <div class="elemHandler">
       <select v-if="projectData.main.data" v-model="curretLevel">
-        <option v-for="(level, i) in projectData.main.data.maxLevel + 1" :key="i">{{ i }}</option>
+        <option
+          v-for="(level, i) in projectData.main.data.maxLevel + 1"
+          :key="i"
+        >
+          {{ i }}
+        </option>
       </select>
       <select v-if="projectData.main.data" v-model="parentSelected">
         <option></option>
-        <option v-for="parent in projectData.main.data[curretLevel]" :key="parent">{{ parent }}</option>
+        <option
+          v-for="parent in projectData.main.data[curretLevel]"
+          :key="parent"
+        >
+          {{ parent }}
+        </option>
       </select>
       <input type="text" v-model="elementToAdd" />
       <button @click="addElement">Add</button>
     </div>
+
+    <Render />
 
     <div>
       <h2>Share this Project with</h2>
@@ -29,7 +41,12 @@
       <div class="overlook">
         <div v-if="parseInt(curretLevel) - 1 >= 0">
           <h3>Previus Level</h3>
-          <p v-for="parent in projectData.main.data[parseInt(curretLevel) - 1]" :key="parent + (curretLevel - 1)">{{ parent }}</p>
+          <p
+            v-for="parent in projectData.main.data[parseInt(curretLevel) - 1]"
+            :key="parent + (curretLevel - 1)"
+          >
+            {{ parent }}
+          </p>
         </div>
         <div v-else>
           <h5>Du bist bereits Toplevel</h5>
@@ -38,7 +55,12 @@
       <div class="overlook">
         <div v-if="projectData.main.data">
           <h3>This Level</h3>
-          <p v-for="parent in projectData.main.data[parseInt(curretLevel)]" :key="parent + curretLevel">{{ parent }}</p>
+          <p
+            v-for="parent in projectData.main.data[parseInt(curretLevel)]"
+            :key="parent + curretLevel"
+          >
+            {{ parent }}
+          </p>
         </div>
         <div v-else>
           <h5>No Data</h5>
@@ -47,7 +69,12 @@
       <div class="overlook">
         <div v-if="projectData.main.data[parseInt(curretLevel) + 1]">
           <h3>Next Level</h3>
-          <p v-for="parent in projectData.main.data[parseInt(curretLevel) + 1]" :key="parent + (curretLevel + 1)">{{ parent }}</p>
+          <p
+            v-for="parent in projectData.main.data[parseInt(curretLevel) + 1]"
+            :key="parent + (curretLevel + 1)"
+          >
+            {{ parent }}
+          </p>
         </div>
         <div v-else>
           <h5>Es gibt keine weiteren Elemnte mehr in dieser reihenfolge</h5>
@@ -62,10 +89,15 @@
 </template>
 
 <script>
+import Render from "../components/Render.vue";
+
 import io from "socket.io-client";
 import api from "../apiService";
 
 export default {
+  components: {
+    Render,
+  },
   data() {
     return {
       ProjectAccessLevel: "",
